requirements.txt
python-telegram-bot==20.8
python-dotenv==0.19.0
firebase-admin==5.3.0
requests==2.28.0
python-dateutil==2.8.2
flask==3.0.0
gunicorn==21.2.0
